+++ java/component/keyple-core/src/main/java/org/eclipse/keyple/core/seproxy/event/AbstractDefaultSelectionsResponse.java
@@ -25,5 +25,5 @@ public abstract class AbstractDefaultSelectionsResponse {
     /**
      * @return the list of {@link SeResponse}
      */
-    public abstract List<SeResponse> getSelectionSeResponseSet();
+    protected abstract List<SeResponse> getSelectionSeResponseSet();
 }
