+++ java/component/keyple-core/src/test/java/org/eclipse/keyple/core/seproxy/plugin/local/AbsObservableLocalReaderTest.java
@@ -84,141 +84,6 @@ public class AbsObservableLocalReaderTest extends CoreBaseTest {
 
 
 
-    /** ==== Card event ====================================== */
-
-    /*
-     * no default selection
-     */
-    @Test
-    public void seInserted() throws Exception {
-        // empty reader
-        AbstractObservableLocalReader r = getSpy(PLUGIN_NAME, READER_NAME);
-
-        // test
-        ReaderEvent event = r.processSeInserted();
-
-        Assert.assertEquals(ReaderEvent.EventType.SE_INSERTED, event.getEventType());
-        Assert.assertTrue(
-                event.getDefaultSelectionsResponse().getSelectionSeResponseSet().isEmpty());
-        Assert.assertEquals(PLUGIN_NAME, event.getPluginName());
-        Assert.assertEquals(READER_NAME, event.getReaderName());
-    }
-
-    /*
-     * selection is not successful
-     */
-    @Test
-    public void seInserted_ALWAYS() throws Exception {
-        AbstractObservableLocalReader r = getSpy(PLUGIN_NAME, READER_NAME);
-
-        // configure parameters
-        Set<SeRequest> selections = new HashSet<SeRequest>();
-        MultiSeRequestProcessing multi = MultiSeRequestProcessing.PROCESS_ALL;
-        ChannelControl channel = ChannelControl.CLOSE_AFTER;
-        ObservableReader.NotificationMode mode = ObservableReader.NotificationMode.ALWAYS;
-
-        // mock return matching selection
-        List<SeResponse> responses = getNotMatchingResponses();
-        doReturn(responses).when(r).transmitSet(selections, multi, channel);
-
-        // test
-        r.setDefaultSelectionRequest(new DefaultSelectionsRequest(selections, multi, channel),
-                mode);
-        ReaderEvent event = r.processSeInserted();
-
-        // assert
-        Assert.assertEquals(ReaderEvent.EventType.SE_INSERTED, event.getEventType());
-        Assert.assertEquals(responses,
-                event.getDefaultSelectionsResponse().getSelectionSeResponseSet());
-        Assert.assertEquals(PLUGIN_NAME, event.getPluginName());
-        Assert.assertEquals(READER_NAME, event.getReaderName());
-
-    }
-
-    /*
-     * selection is not successful
-     */
-    @Test
-    public void noEvent_MATCHED_ONLY() throws Exception {
-        AbstractObservableLocalReader r = getSpy(PLUGIN_NAME, READER_NAME);
-
-        // configure parameters
-        Set<SeRequest> selections = new HashSet<SeRequest>();
-        MultiSeRequestProcessing multi = MultiSeRequestProcessing.PROCESS_ALL;
-        ChannelControl channel = ChannelControl.CLOSE_AFTER;
-        ObservableReader.NotificationMode mode = ObservableReader.NotificationMode.MATCHED_ONLY;
-
-        // mock return matching selection
-        doReturn(getNotMatchingResponses()).when(r).transmitSet(selections, multi, channel);
-
-        // test
-        r.setDefaultSelectionRequest(new DefaultSelectionsRequest(selections, multi, channel),
-                mode);
-        ReaderEvent event = r.processSeInserted();
-
-        Assert.assertEquals(null, event);
-    }
-
-    /*
-     * selection is successful
-     */
-    @Test
-    public void seMatched_MATCHED_ONLY() throws Exception {
-        AbstractObservableLocalReader r = getSpy(PLUGIN_NAME, READER_NAME);
-
-        // configure parameters
-        Set<SeRequest> selections = new HashSet<SeRequest>();
-        MultiSeRequestProcessing multi = MultiSeRequestProcessing.PROCESS_ALL;
-        ChannelControl channel = ChannelControl.CLOSE_AFTER;
-        ObservableReader.NotificationMode mode = ObservableReader.NotificationMode.MATCHED_ONLY;
-
-        // mock
-        // return success selection
-        List<SeResponse> responses = getMatchingResponses();
-        doReturn(responses).when(r).transmitSet(selections, multi, channel);
-
-        // test
-        r.setDefaultSelectionRequest(new DefaultSelectionsRequest(selections, multi, channel),
-                mode);
-        ReaderEvent event = r.processSeInserted();
-
-        Assert.assertEquals(ReaderEvent.EventType.SE_MATCHED, event.getEventType());
-        Assert.assertEquals(responses,
-                event.getDefaultSelectionsResponse().getSelectionSeResponseSet());
-        Assert.assertEquals(PLUGIN_NAME, event.getPluginName());
-        Assert.assertEquals(READER_NAME, event.getReaderName());
-
-    }
-
-
-    /*
-     * Simulate an IOException while selecting Do not throw any event Nor an exception
-     */
-    @Test
-    public void noEvent_IOError() throws Exception {
-        AbstractObservableLocalReader r = getSpy(PLUGIN_NAME, READER_NAME);
-        final CountDownLatch lock = new CountDownLatch(1);
-
-        // configure parameters
-        Set<SeRequest> selections = new HashSet<SeRequest>();
-        MultiSeRequestProcessing multi = MultiSeRequestProcessing.PROCESS_ALL;
-        ChannelControl channel = ChannelControl.CLOSE_AFTER;
-        ObservableReader.NotificationMode mode = ObservableReader.NotificationMode.ALWAYS;
-
-        // throw IO
-        doThrow(new KeypleIOReaderException("io error when selecting")).when(r)
-                .transmitSet(selections, multi, channel);
-
-
-        // test
-        r.setDefaultSelectionRequest(new DefaultSelectionsRequest(selections, multi, channel),
-                mode);
-        r.processSeInserted();
-
-        // test
-        ReaderEvent event = r.processSeInserted();
-        Assert.assertEquals(null, event);
-    }
 
     /**
      * State Machine
